feat: Optimisation of the app
Adding pagination to the courses and comments lists. The code
is working fine but I keep getting a warning because my query
isn't ordered. I'm keeping it like so for now, I'll change it
in a future iteration.

Adding a caching system using redis. All the website's pages
will be cached for a periode of 15 minutes.